+++ src/main/java/br/com/codex/v1/domain/compras/PedidoItensCompra.java
@@ -13,7 +13,7 @@ public class PedidoItensCompra implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Integer id;
+    protected Long id;
     protected String codigoProduto;
     protected String descricaoProduto;
     protected Integer quantidade;
@@ -33,7 +33,7 @@ public class PedidoItensCompra implements Serializable {
         super();
     }
 
-    public PedidoItensCompra(Integer id, String codigoProduto, String descricaoProduto, Integer quantidade, String unidadeComercial, BigDecimal precoUnitario, BigDecimal ipi, BigDecimal pis, BigDecimal icms, BigDecimal cofins, BigDecimal desconto, BigDecimal precoTotal, PedidoCompra pedidoCompra) {
+    public PedidoItensCompra(Long id, String codigoProduto, String descricaoProduto, Integer quantidade, String unidadeComercial, BigDecimal precoUnitario, BigDecimal ipi, BigDecimal pis, BigDecimal icms, BigDecimal cofins, BigDecimal desconto, BigDecimal precoTotal, PedidoCompra pedidoCompra) {
         this.id = id;
         this.codigoProduto = codigoProduto;
         this.descricaoProduto = descricaoProduto;
@@ -49,11 +49,11 @@ public class PedidoItensCompra implements Serializable {
         this.pedidoCompra = pedidoCompra;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
